requirements.txt
# Bot Discord Chronos V2


discord.py==2.3.2


python-dotenv==1.0.0


python-dateutil==2.8.2
